Changed query to only show active tags
If a tag is deleted the current query will get the tag since when a tag is deleted the tag_count table does not have the tag deleted.  

The tag should be deleted from the tag_count & product_tag tables when a tag is deleted in my opinion.  If you turn the product on and off that had the tag then the tag is deleted, I'm not sure if it has to be the product with the tag.